pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sam3-ultralytics"
version = "0.1.0"
description = "API-first SAM 3 wrapper over Ultralytics with a PySide6 desktop GUI and CLI."
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "OpenAI Codex"}]
dependencies = [
  "numpy>=1.26",
  "opencv-python>=4.10",
  "Pillow>=10.0",
  "PySide6>=6.7",
  "timm>=1.0.25",
  "ultralytics>=8.4.24",
  "clip @ git+https://github.com/ultralytics/CLIP.git",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
sam3-gui = "sam3_ultralytics.gui:main"
sam3-ultralytics = "sam3_ultralytics.cli:main"

[tool.setuptools.packages.find]
include = ["sam3_ultralytics*"]